requirements.txt
yt-dlp==2023.7.6
uvicorn[standard]
fastapi==0.101.0
pydantic==2.1.1
pydantic-settings==2.0.2
celery[redis]==5.3.1
requests==2.31.0
websocket-client==1.6.1
python-socketio[client]==5.8.0